+++ test/features/personalization/mocks/actions/manifestReplace.json
@@ -1,0 +1,48 @@
+{
+  "total": 5,
+  "offset": 0,
+  "limit": 5,
+  "data": [
+    {
+      "action": "replace",
+      "selector": ".how-to",
+      "page filter (optional)": "",
+      "param-newoffer=123": "",
+      "chrome": "",
+      "firefox": "/drafts/rwoblesk/personalization-testing/fragments/milo-replace-content-firefox-accordion",
+      "android": "",
+      "ios": ""
+    },
+    {
+      "action": "replace",
+      "selector": "#features-of-milo-experimentation-platform",
+      "page filter (optional)": "",
+      "param-newoffer=123": "",
+      "chrome": "/fragments/milo-replace-content-chrome-howto-h2",
+      "firefox": "",
+      "android": "",
+      "ios": ""
+    },
+    {
+      "action": "replace",
+      "selector": "/fragments/replaceme",
+      "page filter (optional)": "",
+      "param-newoffer=123": "",
+      "chrome": "/fragments/fragmentreplaced",
+      "firefox": "",
+      "android": "",
+      "ios": ""
+    },
+    {
+      "action": "replace",
+      "selector": "/fragments/inline-replaceme",
+      "page filter (optional)": "",
+      "param-newoffer=123": "",
+      "chrome": "/fragments/inline-fragmentreplaced",
+      "firefox": "",
+      "android": "",
+      "ios": ""
+    }
+  ],
+  ":type": "sheet"
+}
