+++ test/features/personalization/mocks/actions/manifestReplacePage.json
