+++ test/features/personalization/mocks/actions/manifestUpdateMetadata.json
