+++ test/features/personalization/mocks/actions/manifestUseBlockCode.json
