+++ test/features/personalization/mocks/actions/manifestUseBlockCode2.json
