+++ test/features/personalization/mocks/deprecatedActions/manifestInsertContentAfter.json
