+++ test/features/personalization/mocks/deprecatedActions/manifestInsertContentBefore.json
