+++ test/features/personalization/mocks/deprecatedActions/manifestRemoveContent.json
