+++ test/features/personalization/mocks/deprecatedActions/manifestReplaceContent.json
