+++ test/features/personalization/mocks/deprecatedActions/manifestReplaceFragment.json
